aggregate logic moved to package, call PurgeAnalyticsDataAggregated added
